Switch order of two items in the document
The "Using the Default Document Root" section comes before "Not Using Standard Document Root Locations".

This can be a little confusing - a 'standard document root location' could be `/var/www/something`, but that could be the default nginx package location.  It is dangerous to use the Default Document Root only because your site is opened to data overwrites on upgrades to the software packages in the given release.  It makes sense to switch the order of these as the "Default Document Root", while possibly a standard docroot, should be specially considered *after* understanding where the standard docroot locations are.